ai-agent-starter-kit/templates/email-assistant/requirements.txt
openai==1.54.0
pyyaml==6.0.2
python-dotenv==1.0.1
